+++ src/bin.rs
@@ -1,31 +1,43 @@
 extern crate rsreddit;
 
 use rsreddit::client::Reddit;
-use rsreddit::model::sort_time::SortTime;
-use rsreddit::oauth2::{AuthorizationTime, RedditApiScope, RedditOAuth};
-use rsreddit::util::convert_scope_vec_to_string;
+use rsreddit::model::listing::Listing;
+use rsreddit::model::listing::ListingCollection;
 
 fn main() {
-    // Show hot posts for authorized user
-    // Works the same for:
-    // * reddit.hot()
-    // * reddit.controversial()
-    // * reddit.new()
-    // * reddit.rising()
-    // * reddit.best()
+    // Get Comment Tree for a thread by its permalink
+    let reddit = Reddit::default().build();
+    let tree = reddit
+        .thread_by_permalink(
+            "/r/rust/comments/hwuvmf/swc_now_works_with_stable_rustc/",
+        )
+        .unwrap();
+    print_comment_tree(&tree);
+}
 
-    // OAuth2 Authorization stuff
-    let mut reddit_oauth = RedditOAuth::default().build();
-    let mut scopes = Vec::new();
-    scopes.push(RedditApiScope::submit);
-    let scope_string = convert_scope_vec_to_string(&scopes);
-    // Authenticate user. Returns bearer token
-    let bearer_token =
-        reddit_oauth.authorize_client(&scope_string, Some(AuthorizationTime::permanent));
-    if let Some(token) = bearer_token {
-        // Reddit client object
-        let reddit = Reddit::default().bearer_token(token).build();
-        // Post comment to toplevel thread
-        reddit.comment("<thing_id>", "Your comment here :)");
-    }
+fn print_comment_tree(tree: &ListingCollection) {
+    //print Thread Title
+    println!(
+        "Title: {}",
+        tree.listings[0].data.children[0]
+            .data
+            .title
+            .as_ref()
+            .unwrap()
+    );
+    //Print comments recursivley
+    travers_comment_tree(&tree.listings[1], 0);
+}
+
+fn travers_comment_tree(comment: &Listing, tab_count: usize) {
+    comment.data.children.iter().for_each(|c| {
+        //Generate right amount if identation
+        let tabs = std::iter::repeat("   ").take(tab_count).collect::<String>();
+        //Print comment
+        println!("{}>{}", tabs, c.data.body.as_ref().unwrap());
+        //Go a level deeper
+        if let Some(replies) = &c.data.replies {
+            travers_comment_tree(replies,tab_count + 1);
+        }
+    });
 }